# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AgriChain Digital Agriculture Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List farmer groups
      parameters:
        - in: query
          name: regionId
          schema:
            type: string
      responses:
        "200":
          description: Group list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FarmerGroup"

    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new farmer group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupBody"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmerGroup"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group details including membership
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmerGroupDetail"

  # ============================================================
  # PLOTS
  # ============================================================
  /plots:
    get:
      operationId: listPlots
      tags: [plots]
      summary: List farm plots
      parameters:
        - in: query
          name: farmerId
          schema:
            type: string
      responses:
        "200":
          description: Plot list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plot"

    post:
      operationId: createPlot
      tags: [plots]
      summary: Create a farm plot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlotBody"
      responses:
        "201":
          description: Plot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plot"
  /plots/{plotId}:
    get:
      operationId: getPlot
      tags: [plots]
      summary: Get plot by ID
      parameters:
        - in: path
          name: plotId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plot"

  # ============================================================
  # CERTIFICATIONS
  # ============================================================
...
  /surveys/templates:
    get:
      operationId: listSurveyTemplates
      tags: [surveys]
      summary: List survey templates
      responses:
        "200":
          description: Survey templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SurveyTemplate"

    post:
      operationId: createSurveyTemplate
      tags: [surveys]
      summary: Create a survey template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyTemplateBody"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyTemplate"
  /surveys/submissions:
    get:
      operationId: listSurveySubmissions
      tags: [surveys]
      summary: List submitted surveys (QA dashboard)
      parameters:
        - in: query
          name: agentId
          schema:
            type: string
        - in: query
          name: groupId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected, flagged]
        - in: query
          name: dateFrom
          schema:
            type: string
            format: date
        - in: query
          name: dateTo
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Survey submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SurveySubmission"

    post:
      operationId: submitSurvey
      tags: [surveys]
      summary: Submit a completed survey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitSurveyBody"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveySubmission"
  /surveys/submissions/{submissionId}/review:
    post:
      operationId: reviewSurveySubmission
      tags: [surveys]
      summary: Approve, reject or flag a survey submission
      parameters:
        - in: path
          name: submissionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewSubmissionBody"
      responses:
        "200":
          description: Updated submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveySubmission"

  # ============================================================
  # VISITS
  # ============================================================
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List farmer payments
      parameters:
        - in: query
          name: farmerId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, processing, paid, failed]
        - in: query
          name: dateFrom
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate a farmer payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiatePaymentBody"
      responses:
        "201":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List staff members
      parameters:
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: regionId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, on_leave, deactivated]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createUser
      tags: [users]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ============================================================
  # ASSETS
  # ============================================================
...
components:
  schemas:
...
    Farmer:
      type: object
      properties:
        id:
          type: string
        referenceNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        nationalId:
          type: string
        dateOfBirth:
          type: string
          format: date
        phoneNumber:
          type: string
        sex:
          type: string
          enum: [male, female, other]
        groupId:
          type: string
        groupName:
          type: string
        regionId:
          type: string
        village:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
        photoUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, referenceNumber, firstName, lastName, groupId, regionId, status, createdAt, updatedAt]
...
    FarmerGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        regionId:
          type: string
        village:
          type: string
        memberCount:
          type: integer
        complianceScore:
          type: number
        activePlots:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, regionId, memberCount, createdAt]
    FarmerGroupDetail:
      allOf:
        - $ref: "#/components/schemas/FarmerGroup"
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/Farmer"
            leaders:
              type: array
              items:
                type: object
                properties:
                  farmerId:
                    type: string
                  name:
                    type: string
                  role:
                    type: string
                  startDate:
                    type: string
                    format: date
                required: [farmerId, name, role, startDate]
          required: [members, leaders]
    CreateGroupBody:
      type: object
      properties:
        name:
          type: string
        regionId:
          type: string
        village:
          type: string
      required: [name, regionId]

    # ============================================================
    # PLOTS
    # ============================================================
    Plot:
      type: object
      properties:
        id:
          type: string
        farmerId:
          type: string
        name:
          type: string
        cropType:
          type: string
        areaHectares:
          type: number
        polygon:
          type: object
          description: GeoJSON polygon
        harvestDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
      required: [id, farmerId, cropType, areaHectares, status, createdAt]
    CreatePlotBody:
      type: object
      properties:
        farmerId:
          type: string
        name:
          type: string
        cropType:
          type: string
        polygon:
          type: object
        harvestDate:
          type: string
          format: date
      required: [farmerId, cropType]

    # ============================================================
    # CERTIFICATIONS
    # ============================================================
...
    SurveyTemplate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [farm_visit, gap_assessment, training_attendance, commodity_inspection, custom]
        fields:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              label:
                type: string
              type:
                type: string
                enum: [text, number, date, dropdown, photo, gps, boolean]
              required:
                type: boolean
              options:
                type: array
                items:
                  type: string
            required: [id, label, type, required]
        published:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, fields, published, createdAt]
    CreateSurveyTemplateBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [farm_visit, gap_assessment, training_attendance, commodity_inspection, custom]
        fields:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              type:
                type: string
              required:
                type: boolean
              options:
                type: array
                items:
                  type: string
            required: [label, type, required]
      required: [name, type, fields]
    SurveySubmission:
      type: object
      properties:
        id:
          type: string
        templateId:
          type: string
        templateName:
          type: string
        farmerId:
          type: string
        farmerName:
          type: string
        agentId:
          type: string
        agentName:
          type: string
        responses:
          type: object
          additionalProperties: true
        gpsLat:
          type: number
        gpsLng:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, flagged]
        reviewComment:
          type: string
        submittedAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
      required: [id, templateId, farmerId, agentId, status, submittedAt]
    SubmitSurveyBody:
      type: object
      properties:
        templateId:
          type: string
        farmerId:
          type: string
        responses:
          type: object
          additionalProperties: true
        gpsLat:
          type: number
        gpsLng:
          type: number
      required: [templateId, farmerId, responses]
    ReviewSubmissionBody:
      type: object
      properties:
        decision:
          type: string
          enum: [approved, rejected, flagged]
        comment:
          type: string
      required: [decision, comment]

    # ============================================================
    # VISITS
    # ============================================================
...
    Payment:
      type: object
      properties:
        id:
          type: string
        farmerId:
          type: string
        farmerName:
          type: string
        deliveryId:
          type: string
        lotTag:
          type: string
        amountDue:
          type: number
        amountPaid:
          type: number
        currency:
          type: string
        paymentMethod:
          type: string
          enum: [mobile_money, bank_transfer, cash]
        status:
          type: string
          enum: [pending, processing, paid, failed]
        paymentReference:
          type: string
        paidAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, farmerId, deliveryId, amountDue, currency, paymentMethod, status, createdAt]
    InitiatePaymentBody:
      type: object
      properties:
        deliveryId:
          type: string
        farmerId:
          type: string
        amountDue:
          type: number
        paymentMethod:
          type: string
          enum: [mobile_money, bank_transfer, cash]
        currency:
          type: string
      required: [deliveryId, farmerId, amountDue, paymentMethod, currency]
...
    User:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        role:
          type: string
          enum: [Agronomist, Farmer, Manager, Supervisor, GroupLeader, BuyingStationAgent, QualityInspector, TradeDesk, ProcurementHead, FinanceOfficer, SystemAdministrator, WarehouseManager, LogisticsOfficer, ComplianceOfficer]
        regionId:
          type: string
        managerId:
          type: string
        status:
          type: string
          enum: [active, on_leave, deactivated]
        createdAt:
          type: string
          format: date-time
      required: [id, firstName, lastName, email, role, status, createdAt]
    CreateUserBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        role:
          type: string
        regionId:
          type: string
        managerId:
          type: string
      required: [firstName, lastName, email, role]
    UpdateUserBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        role:
          type: string
        regionId:
          type: string
        managerId:
          type: string
        status:
          type: string
          enum: [active, on_leave, deactivated]

    # ============================================================
    # ASSETS
    # ============================================================
...